requirements.txt
streamlit
opencv-python
mediapipe
torch
torchvision
torchaudio
pillow
numpy
requests

